Struct rustup::cli::download_tracker::DownloadTracker

pub struct DownloadTracker {
    content_len: Option<usize>,
    total_downloaded: usize,
    downloaded_this_sec: usize,
    downloaded_last_few_secs: VecDeque<usize>,
    last_sec: Option<Instant>,
    start_sec: Option<Instant>,
    term: AutomationFriendlyTerminal<Box<dyn Writer>>,
    displayed_charcount: Option<usize>,
    units: Vec<Unit>,
    display_progress: bool,
}

Tracks download progress and displays information about it to a terminal.

Fields
content_len: Option<usize>

Content-Length of the to-be downloaded object.

total_downloaded: usize

Total data downloaded in bytes.

downloaded_this_sec: usize

Data downloaded this second.

downloaded_last_few_secs: VecDeque<usize>

Keeps track of amount of data downloaded every last few secs. Used for averaging the download speed. NB: This does not necessarily represent adjacent seconds; thus it may not show the average at all.

last_sec: Option<Instant>

Time stamp of the last second

start_sec: Option<Instant>

Time stamp of the start of the download

term: AutomationFriendlyTerminal<Box<dyn Writer>>

The terminal we write the information to. XXX: Could be a term trait, but with #1818 on the horizon that is a pointless change to make - better to let that transition happen and take stock after that.

displayed_charcount: Option<usize>

Whether we displayed progress for the download or not.

If the download is quick enough, we don't have time to display the progress info. In that case, we do not want to do some cleanup stuff we normally do.

If we have displayed progress, this is the number of characters we rendered, so we can erase it cleanly.

units: Vec<Unit>

What units to show progress in

display_progress: bool

Whether we display progress

Implementations

impl DownloadTracker

pub fn new() -> Self

Creates a new DownloadTracker.

pub fn with_display_progress(self, display_progress: bool) -> Self

pub fn handle_notification(&mut self, n: &Notification<'_>) -> bool

pub fn content_length_received(&mut self, content_len: u64)

Notifies self that Content-Length information has been received.

pub fn data_received(&mut self, len: usize)

Notifies self that data of size len has been received.

pub fn download_finished(&mut self)

Notifies self that the download has finished.

fn prepare_for_new_download(&mut self)

Resets the state to be ready for a new download.

fn display(&mut self)

Display the tracked download information to the terminal.

pub fn push_unit(&mut self, new_unit: Unit)

pub fn pop_unit(&mut self)