Module rustup::cli::self_update

Self-installation and updating

This is the installer at the heart of Rust. If it breaks everything breaks. It is conceptually very simple, as rustup is distributed as a single binary, and installation mostly requires copying it into place. There are some tricky bits though, mostly because of workarounds to self-delete an exe on Windows.

During install (as rustup-init):

copy the self exe to $CARGO_HOME/bin

hardlink rustc, etc to that

update the PATH in a system-specific way

run the equivalent of rustup default stable

During upgrade (rustup self upgrade):

download rustup-init to $CARGO_HOME/bin/rustup-init

run rustup-init with appropriate flags to indicate this is a self-upgrade

rustup-init copies bins and hardlinks into place. On windows this happens after the upgrade command exits successfully.

During uninstall (rustup self uninstall):

Delete $RUSTUP_HOME.

Delete everything in $CARGO_HOME, including the rustup binary and its hardlinks

Deleting the running binary during uninstall is tricky and racy on Windows.

Modules

os

shell
Paths and Unix shells

test
Support for functional tests.

unix

Structs

InstallOpts

Constants

NEVER_SELF_UPDATE

Statics

UPDATE_ROOT

Functions

_install_selection

canonical_cargo_home
CARGO_HOME suitable for display, possibly with $HOME substituted for the directory prefix

check_existence_of_rustc_or_cargo_in_path

cleanup_self_updater

current_install_opts

customize_install

delete_rustup_and_cargo_home

do_pre_install_options_sanity_checks

do_pre_install_sanity_checks

get_new_rustup_version

install
Installing is a simple matter of copying the running binary to CARGO_HOME/bin, hard-linking the various Rust tools to it, and adding CARGO_HOME/bin to PATH.

install_bins

install_proxies

maybe_install_rust

parse_new_rustup_version

pre_install_msg

prepare_update

run_update
Tell the upgrader to replace the rustup bins, then delete itself. Like with uninstallation, on Windows we're going to have to jump through hoops to make everything work right.

rustc_or_cargo_exists_in_path

self_replace
This function is as the final step of a self-upgrade. It replaces CARGO_HOME/bin/rustup with the running exe, and updates the the links to it. On windows this will run after the original rustup process exits.

uninstall

update
Self update downloads rustup-init to CARGO_HOME/bin/rustup-init and runs it.