Module rustup::cli::self_update::shell

Paths and Unix shells

MacOS, Linux, FreeBSD, and many other OS model their design on Unix, so handling them is relatively consistent. But only relatively. POSIX postdates Unix by 20 years, and each "Unix-like" shell develops unique quirks over time.

Windowing Managers, Desktop Environments, GUI Terminals, and PATHs

Duplicating paths in PATH can cause performance issues when the OS searches the same place multiple times. Traditionally, Unix configurations have resolved this by setting up PATHs in the shell's login profile.

This has its own issues. Login profiles are only intended to run once, but changing the PATH is common enough that people may run it twice. Desktop environments often choose to NOT start login shells in GUI terminals. Thus, a trend has emerged to place PATH updates in other run-commands (rc) files, leaving Rustup with few assumptions to build on for fulfilling its promise to set up PATH appropriately.

Rustup addresses this by:

using a shell script that updates PATH if the path is not in PATH

sourcing this script in any known and appropriate rc file

Structs

Bash

Posix

ShellScript

Zsh

Traits

UnixShell

Functions

cargo_home_str

enumerate_shells

get_available_shells

legacy_paths

Type Definitions

Shell