Trait rustup::currentprocess::HomeProcess

pub trait HomeProcess: CurrentProcess + Env {
    fn clone_boxed(&self) -> Box<dyn HomeProcess>;
}

An abstraction for the current process

This provides replacements env::arg*, env::var*, and the standard files io::std* with traits that are customisable for tests. As a result any macros or code that have non-pluggable usage of those are incompatible with CurrentProcess and must not be used. That includes [e]println! as well as third party crates.

CurrentProcess is used via an instance in a thread local variable; when making new threads, be sure to copy CurrentProcess::process() into the new thread before calling any code that may need to use a CurrentProcess function.

Run some code using with: this will set the current instance, call your function, then finally reset the instance at the end before returning.

Testing level interoperation with external code that depends on environment variables could be possible with a hypothetical with_projected() which would be a zero-cost operation in real processes, but in test processes will take a lock out to mutually exclude other code, then overwrite the current value of std::env::vars, restoring it at the end. However, the only use for that today is a test of cargo::home, which is now implemented in a separate crate, so we've just deleted the test.

A thread local is used to permit the instance to be available to the entire rustup library without needing to explicitly wire this normally global state everywhere; and a trait object with dyn dispatch is likewise used to avoid needing to thread trait parameters across the entire code base: none of the methods are in performance critical loops (except perhaps progress bars - and even there we should be doing debouncing and managing update rates). The real trait is CurrentProcess; HomeProcess is a single trait because Box only allows autotraits to be added to it; so we use a subtrait to add home::Env in.

Required methods

fn clone_boxed(&self) -> Box<dyn HomeProcess>

Loading content...