Module rustup::diskio

Modules

immediate
Disk IO abstraction for rustup.

threaded

Structs

Item

Enums

Kind

Traits

Executor
Trait object for performing IO. At this point the overhead of trait invocation is not a bottleneck, but if it becomes one we could consider an enum variant based approach instead.

Functions

create_dir

get_executor
Get the executor for disk IO.

perform
Trivial single threaded IO to be used from executors. (Crazy sophisticated ones can obviously ignore this)

write_file