Trait rustup::diskio::Executor

pub trait Executor {
    fn dispatch(&mut self, item: Item) -> Box<dyn Iterator<Item = Item>>;

    fn join(&mut self) -> Box<dyn Iterator<Item = Item>>;

    fn completed(&mut self) -> Box<dyn Iterator<Item = Item>>;

    fn execute(&mut self, item: Item) -> Box<dyn Iterator<Item = Item>> { ... }
}

Trait object for performing IO. At this point the overhead of trait invocation is not a bottleneck, but if it becomes one we could consider an enum variant based approach instead.

Required methods

fn dispatch(&mut self, item: Item) -> Box<dyn Iterator<Item = Item>>

Actually dispatch a operation. This is called by the default execute() implementation and should not be called directly.

fn join(&mut self) -> Box<dyn Iterator<Item = Item>>

Wrap up any pending operations and iterate over them. All operations submitted before the join will have been returned either through ready/complete or join once join returns.

fn completed(&mut self) -> Box<dyn Iterator<Item = Item>>

Iterate over completed items.

Loading content...
Provided methods

fn execute(&mut self, item: Item) -> Box<dyn Iterator<Item = Item>>

Perform a single operation. During overload situations previously queued items may need to be completed before the item is accepted: consume the returned iterator.

Loading content...