Module rustup::dist::component::package

An interpreter for the rust-installer package format. Responsible for installing from a directory or tarball to an installation prefix, represented by a Components instance.

Structs

DirectoryPackage

MemoryBudget

TarGzPackage

TarPackage

TarXzPackage

Enums

DirStatus
What is the status of this directory ?

Constants

INSTALLER_VERSION
The current metadata revision used by rust-installer

VERSION_FILE

Traits

Package

Functions

filter_result
Handle the async result of io operations Replaces op.result with Ok(())

trigger_children
Dequeue the children of directories queued up waiting for the directory to be created.

unpack_without_first_dir

validate_installer_version