Enum rustup::dist::component::transaction::ChangedItem

enum ChangedItem<'a> {
    AddedFile(PathBuf),
    AddedDir(PathBuf),
    RemovedFile(PathBuf, File<'a>),
    RemovedDir(PathBuf, Dir<'a>),
    ModifiedFile(PathBuf, Option<File<'a>>),
}

This is the set of fundamental operations supported on a Transaction. More complicated operations, such as installing a package, or updating a component, distill down into a series of these primitives.

Variants

AddedFile(PathBuf)

AddedDir(PathBuf)

RemovedFile(PathBuf, File<'a>)

RemovedDir(PathBuf, Dir<'a>)

ModifiedFile(PathBuf, Option<File<'a>>)

Implementations

impl<'a> ChangedItem<'a>

fn roll_back(
&self,
prefix: &InstallPrefix,
notify: &'a dyn Fn(Notification<'_>)
) -> Result<()>

fn dest_abs_path(
prefix: &InstallPrefix,
component: &str,
relpath: &PathBuf
) -> Result<PathBuf>

fn add_file(
prefix: &InstallPrefix,
component: &str,
relpath: PathBuf
) -> Result<(Self, File)>

fn copy_file(
prefix: &InstallPrefix,
component: &str,
relpath: PathBuf,
src: &Path
) -> Result<Self>

fn copy_dir(
prefix: &InstallPrefix,
component: &str,
relpath: PathBuf,
src: &Path
) -> Result<Self>

fn remove_file(
prefix: &InstallPrefix,
component: &str,
relpath: PathBuf,
temp_cfg: &'a Cfg,
notify: &'a dyn Fn(Notification<'_>)
) -> Result<Self>

fn remove_dir(
prefix: &InstallPrefix,
component: &str,
relpath: PathBuf,
temp_cfg: &'a Cfg,
notify: &'a dyn Fn(Notification<'_>)
) -> Result<Self>

fn modify_file(
prefix: &InstallPrefix,
relpath: PathBuf,
temp_cfg: &'a Cfg
) -> Result<Self>

fn move_file(
prefix: &InstallPrefix,
component: &str,
relpath: PathBuf,
src: &Path,
notify: &'a dyn Fn(Notification<'_>)
) -> Result<Self>

fn move_dir(
prefix: &InstallPrefix,
component: &str,
relpath: PathBuf,
src: &Path,
notify: &'a dyn Fn(Notification<'_>)
) -> Result<Self>