Struct rustup::dist::component::transaction::Transaction

pub struct Transaction<'a> {
    prefix: InstallPrefix,
    changes: Vec<ChangedItem<'a>>,
    temp_cfg: &'a Cfg,
    notify_handler: &'a dyn Fn(Notification<'_>),
    committed: bool,
}

A Transaction tracks changes to the file system, allowing them to be rolled back in case of an error. Instead of deleting or overwriting file, the old copies are moved to a temporary folder. If the transaction is rolled back, they will be moved back into place. If the transaction is committed, these files are automatically cleaned up using the temp system.

All operations that create files will automatically create any intermediate directories in the path to the file if they do not already exist.

All operations that create files will fail if the destination already exists.

Fields
prefix: InstallPrefixchanges: Vec<ChangedItem<'a>>temp_cfg: &'a Cfgnotify_handler: &'a dyn Fn(Notification<'_>)committed: bool
Implementations

impl<'a> Transaction<'a>

pub fn new(
prefix: InstallPrefix,
temp_cfg: &'a Cfg,
notify_handler: &'a dyn Fn(Notification<'_>)
) -> Self

pub fn commit(self)

Commit must be called for all successful transactions. If not called the transaction will be rolled back on drop.

fn change(&mut self, item: ChangedItem<'a>)

pub fn add_file(&mut self, component: &str, relpath: PathBuf) -> Result<File>

Add a file at a relative path to the install prefix. Returns a File that may be used to subsequently write the contents.

pub fn copy_file(
&mut self,
component: &str,
relpath: PathBuf,
src: &Path
) -> Result<()>

Copy a file to a relative path of the install prefix.

pub fn copy_dir(
&mut self,
component: &str,
relpath: PathBuf,
src: &Path
) -> Result<()>

Recursively copy a directory to a relative path of the install prefix.

pub fn remove_file(&mut self, component: &str, relpath: PathBuf) -> Result<()>

Remove a file from a relative path to the install prefix.

pub fn remove_dir(&mut self, component: &str, relpath: PathBuf) -> Result<()>

Recursively remove a directory from a relative path of the install prefix.

pub fn write_file(
&mut self,
component: &str,
relpath: PathBuf,
content: String
) -> Result<()>

Create a new file with string contents at a relative path to the install prefix.

pub fn modify_file(&mut self, relpath: PathBuf) -> Result<()>

If the file exists back it up for rollback, otherwise ensure that the path to it exists so that subsequent calls to File::create will succeed.

This is used for arbitrarily manipulating a file.

pub fn move_file(
&mut self,
component: &str,
relpath: PathBuf,
src: &Path
) -> Result<()>

Move a file to a relative path of the install prefix.

pub fn move_dir(
&mut self,
component: &str,
relpath: PathBuf,
src: &Path
) -> Result<()>

Recursively move a directory to a relative path of the install prefix.

pub fn temp(&self) -> &'a Cfg

pub fn notify_handler(&self) -> &'a dyn Fn(Notification<'_>)