Struct rustup::dist::manifestation::Update

struct Update {
    components_to_uninstall: Vec<Component>,
    components_to_install: Vec<Component>,
    final_component_list: Vec<Component>,
    missing_components: Vec<Component>,
}

Fields
components_to_uninstall: Vec<Component>components_to_install: Vec<Component>final_component_list: Vec<Component>missing_components: Vec<Component>
Implementations

impl Update

fn build_update(
manifestation: &Manifestation,
new_manifest: &Manifest,
changes: &Changes,
config: &Option<Config>,
notify_handler: &dyn Fn(Notification<'_>)
) -> Result<Self>

Returns components to uninstall, install, and the list of all components that will be up to date after the update.

fn build_final_component_list(
&mut self,
starting_list: &[Component],
rust_target_package: &TargetedPackage,
new_manifest: &Manifest,
changes: &Changes
)

Build the list of components we'll have installed at the end

fn nothing_changes(&self) -> bool

fn unavailable_components(
&self,
new_manifest: &Manifest,
toolchain_str: &str
) -> Result<()>

fn drop_components_to_install(&mut self, to_drop: &[Component])

fn components_urls_and_hashes(
&self,
new_manifest: &Manifest
) -> Result<Vec<(Component, Format, String, String)>>

Map components to urls and hashes