Module rustup::test

Test support module; public to permit use from integration tests.

Re-exports

pub use crate::cli::self_update::test::get_path;

pub use crate::cli::self_update::test::with_saved_path;

Structs

RustupHome
The smallest form of test isolation: an isolated RUSTUP_HOME, for codepaths that read and write config files but do not invoke processes, download data etc.

Traits

Env

Functions

tempdir_in_with_prefix
Makes persistent unique directory inside path.

test_dir
Returns a tempdir for running tests in

this_host_triple
What is this host's triple - seems very redunant with from_host_or_build() ... perhaps this is so that the test data we have is only exercised on known triples?

with_rustup_home
Create an isolated rustup home with no content, then call f with it, and delete it afterwards.