Enum rustup::utils::notifications::Notification

pub enum Notification<'a> {
    CreatingDirectory(&'a str, &'a Path),
    LinkingDirectory(&'a Path, &'a Path),
    CopyingDirectory(&'a Path, &'a Path),
    RemovingDirectory(&'a str, &'a Path),
    DownloadingFile(&'a Url, &'a Path),
    DownloadContentLengthReceived(u64),
    DownloadDataReceived(&'a [u8]),
    DownloadFinished,
    DownloadPushUnit(Unit),
    DownloadPopUnit,
    NoCanonicalPath(&'a Path),
    ResumingPartialDownload,
    SetDefaultBufferSize(usize),
    Error(String),
    UsingCurl,
    UsingReqwest,
    RenameInUse(&'a Path, &'a Path),
}

Variants

CreatingDirectory(&'a str, &'a Path)

LinkingDirectory(&'a Path, &'a Path)

CopyingDirectory(&'a Path, &'a Path)

RemovingDirectory(&'a str, &'a Path)

DownloadingFile(&'a Url, &'a Path)

DownloadContentLengthReceived(u64)

Received the Content-Length of the to-be downloaded data.

DownloadDataReceived(&'a [u8])

Received some data.

DownloadFinished

Download has finished.

DownloadPushUnit(Unit)

The things we're tracking that are not counted in bytes. Must be paired with a pop-units; our other calls are not setup to guarantee this any better.

DownloadPopUnit

finish using an unusual unit.

NoCanonicalPath(&'a Path)

ResumingPartialDownload

SetDefaultBufferSize(usize)

This would make more sense as a crate::notifications::Notification member, but the notification callback is already narrowed to utils::notifications by the time tar unpacking is called.

Error(String)

UsingCurl

UsingReqwest

RenameInUse(&'a Path, &'a Path)

Renaming encountered a file in use error and is retrying. The InUse aspect is a heuristic - the OS specifies Permission denied, but as we work in users home dirs and running programs like virus scanner are known to cause this the heuristic is quite good.

Implementations

impl<'a> Notification<'a>

pub fn level(&self) -> NotificationLevel